Change method used to convert encoding name to C string
The libxml call to parse HTML takes a C string to designate the encoding. This
is derived from the NSUTF8StringEncoding constant through a couple
CFString function calls. However, in the last step, the function
`CFStringGetCStringPtr` was used to convert the encoding name CFString to a
C string. The implementation details of this function apparently
changed from iOS8 to iOS9, so it was returning NULL in iOS9 instead of
"utf-8". I guess this is why Apple warns not to use it. (http://stackoverflow.c
om/questions/18976485/cfstringgetcstringptr-returns-null-on-ios7) We use the
more stable `-NSString UTF8String` instead.